+++ admin/themes/default/index.template.php
@@ -2,12 +2,12 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    
+
     <meta http-equiv="x-dns-prefetch-control" content="on">
     <link rel="dns-prefetch" href="<?php echo Site::url(); ?>" />
     <link rel="dns-prefetch" href="//www.google-analytics.com" />
     <link rel="dns-prefetch" href="//www.gravatar.com" />
-    
+
     <title>Monstra :: <?php echo __('Administration', 'system'); ?></title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Monstra Admin Area" />
@@ -27,8 +27,8 @@
     <!-- JavaScripts -->
     <script src="<?php echo Site::url(); ?>/public/assets/js/jquery.min.js"></script>
     <script src="<?php echo Site::url(); ?>/public/assets/js/bootstrap.min.js"></script>
-    <script src="<?php echo Site::url(); ?>/public/assets/js/messenger.min.js"></script>    
-    <script src="<?php echo Site::url(); ?>/public/assets/js/icheck.min.js"></script>    
+    <script src="<?php echo Site::url(); ?>/public/assets/js/messenger.min.js"></script>
+    <script src="<?php echo Site::url(); ?>/public/assets/js/icheck.min.js"></script>
     <?php Javascript::add('public/assets/js/jquery.chocolat.js', 'backend', 3); ?>
     <?php Javascript::add('public/assets/js/bootstrap-fileupload.js', 'backend', 4); ?>
     <?php Javascript::add('admin/themes/default/js/default.js', 'backend', 5); ?>
@@ -41,9 +41,9 @@
 
           $('.chocolat').Chocolat({
               overlayColor          : '#000',
-              leftImg               : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/left.gif",  
-              rightImg              : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/right.gif", 
-              closeImg              : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/close.gif",   
+              leftImg               : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/left.gif",
+              rightImg              : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/right.gif",
+              closeImg              : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/close.gif",
               loadingImg            : "<?php echo Option::get('siteurl'); ?>/public/assets/img/chocolat/loading.gif"
           });
 
@@ -53,7 +53,7 @@
               increaseArea: '20%'
           });
 
-        });      
+        });
     </script>
 
     <!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
@@ -65,7 +65,7 @@
 
   <body class="page-<?php echo Request::get('id'); ?>">
 
-    <nav class="navbar navbar-default navbar-inverse" role="navigation">      
+    <nav class="navbar navbar-default navbar-inverse" role="navigation">
       <div class="container">
           <div class="navbar-header">
             <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
@@ -77,33 +77,33 @@
             <a class="navbar-brand" href="<?php echo Site::url(); ?>/admin/index.php?id=dashboard">MONSTRA</a>
           </div>
 
-          <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">      
-            <ul class="nav navbar-nav">          
-              <li<?php if (Request::get('id') == 'dashboard') { ?> class="active"<?php } ?>><a href="<?php echo Site::url(); ?>/admin/index.php?id=dashboard"><?php echo __('Dashboard', 'dashboard'); ?></a></li>              
+          <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+            <ul class="nav navbar-nav">
+              <li<?php if (Request::get('id') == 'dashboard') { ?> class="active"<?php } ?>><a href="<?php echo Site::url(); ?>/admin/index.php?id=dashboard"><?php echo __('Dashboard', 'dashboard'); ?></a></li>
               <li class="dropdown">
                 <a href="#" class="dropdown-toggle" data-toggle="dropdown"><?php echo __('Content', 'pages'); ?> <b class="caret"></b></a>
                 <ul class="dropdown-menu">
                     <?php Navigation::draw('content'); ?>
                 </ul>
               </li>
-              <?php if (Session::exists('user_role') && in_array(Session::get('user_role'), array('admin'))) { ?>               
+              <?php if (Session::exists('user_role') && in_array(Session::get('user_role'), array('admin'))) { ?>
               <li class="dropdown">
                 <a href="#" class="dropdown-toggle" data-toggle="dropdown"><?php echo __('Extends', 'system'); ?> <b class="caret"></b></a>
                 <ul class="dropdown-menu">
-                    <?php Navigation::draw('extends'); ?>                      
+                    <?php Navigation::draw('extends'); ?>
                 </ul>
               </li>
               <?php } ?>
               <li class="dropdown">
                 <a href="#" class="dropdown-toggle" data-toggle="dropdown"><?php echo __('System', 'system'); ?> <b class="caret"></b></a>
                 <ul class="dropdown-menu">
-                    <?php Navigation::draw('system'); ?>                       
+                    <?php Navigation::draw('system'); ?>
                 </ul>
               </li>
               <li class="dropdown">
                 <a href="#" class="dropdown-toggle" data-toggle="dropdown"><?php echo __('Help', 'system'); ?> <b class="caret"></b></a>
                 <ul class="dropdown-menu">
-                    <li><a href="http://monstra.org/documentation" target="_blank"><?php echo __('Documentation', 'system'); ?></a></li>              
+                    <li><a href="http://monstra.org/documentation" target="_blank"><?php echo __('Documentation', 'system'); ?></a></li>
                     <?php if (Option::get('language') == 'ru') { ?>
                     <li><a href="http://ru.forum.monstra.org" target="_blank"><?php echo __('Official Support Forum', 'system'); ?></a></li>
                     <?php } else { ?>
@@ -111,23 +111,23 @@
                     <?php } ?>
                 </ul>
               </li>
-            </ul>        
+            </ul>
             <ul class="nav navbar-nav navbar-right">
               <li><a href="<?php echo Site::url(); ?>" target="_blank"><?php echo __('View Site', 'system'); ?></a></li>
               <li class="dropdown">
                 <a href="#" class="dropdown-toggle" data-toggle="dropdown"><?php echo Session::get('user_login'); ?> <img src="<?php echo Users::getGravatarURL(Session::get('user_email'), 28); ?>" alt=""> <b class="caret"></b></a>
                 <ul class="dropdown-menu">
                   <li><a href="<?php echo Site::url(); ?>/admin/index.php?id=users&action=edit&user_id=<?php echo Session::get('user_id'); ?>"><?php echo __('Profile', 'users')?></a></li>
-                  <li><a href="<?php echo Site::url(); ?>/admin/?logout=do"><?php echo __('Log Out', 'users'); ?></a></li>              
+                  <li><a href="<?php echo Site::url(); ?>/admin/?logout=do"><?php echo __('Log Out', 'users'); ?></a></li>
                 </ul>
               </li>
-            </ul>        
+            </ul>
           </div>
       </div>
     </nav>
 
     <div class="container">
-        
+
         <?php
             // Monstra Notifications
             Notification::get('success') AND Alert::success(Notification::get('success'));
@@ -156,13 +156,7 @@
       <footer class="container visible-md visible-lg">
           <p class="pull-right">
             <span>
-              <?php if (Option::get('language') == 'ru') { ?>
-              <a href="http://ru.forum.monstra.org" target="_blank"><?php echo __('Official Support Forum', 'system'); ?></a> /
-              <?php } else { ?>
-              <a href="http://forum.monstra.org" target="_blank"><?php echo __('Official Support Forum', 'system'); ?></a> /
-              <?php } ?>
-              <a href="http://monstra.org/documentation" target="_blank"><?php echo __('Documentation', 'system'); ?></a> /
-              © 2012 - 2014 <a href="http://monstra.org" target="_blank">Monstra</a> – <?php echo __('Version', 'system'); ?> <?php echo Monstra::VERSION; ?>
+              <a href="http://monstra.org" target="_blank">Monstra</a> was made by <a href="http://awilum.github.io" target="_blank" class="highlight">Sergey Romanenko</a> and is maintained by <a href="https://github.com/monstra-cms/monstra/network/members" target="_blank" class="highlight">Monstra Community</a> / © 2012 - 2016 <a href="http://monstra.org/about/license" target="_blank">Monstra</a> – <?php echo __('Version', 'system'); ?> <?php echo Monstra::VERSION; ?>
             </span>
           </p>
       </footer>