+++ engine/Monstra.php
@@ -31,7 +31,7 @@ class Monstra
     /**
      * The version of Monstra
      */
-    const VERSION = '3.0.2';
+    const VERSION = '3.0.3';
 
 
     /**