+++ plugins/box/editor/languages/ka-ge.lang.php
@@ -1,0 +1,7 @@
+<?php
+    return array(
+        'Editor' => array(
+            'Editor' => 'რედაქტორი',
+            'Editor plugin' => 'პლაგინი რედაქტორი',
+        )
+    );