+++ plugins/box/information/languages/ka-ge.lang.php
@@ -1,0 +1,40 @@
+<?php
+
+    return array(
+        'information' => array(
+            'Information' => 'ინფორმაცია',
+            'Debugging' => 'Debugging-ი',
+            'Name' => 'სახელი',
+            'Value' => 'მნიშვნელობა',
+            'Security' => 'უსაფრთხოება',
+            'System' => 'სისტემა',
+            'on' => 'ჩართულია',
+            'off'=> 'გამორთულია',
+            'Server' => 'სერვერი',
+            'PHP version' => 'PHP-ის ვერსია',
+            'SimpleXML module' => 'SimpleXML მოდული',
+            'DOM module' => 'DOM მოდული',
+            'Installed' => 'დაყენებულია',
+            'Not Installed' => 'არ არის დაყენებული',
+            'Security check results' => 'უსაფრთხოების შემოწმების შედეგები',
+            'The configuration file has been found to be writable. We would advise you to remove all write permissions on defines.php on production systems.' =>
+            'კონფიგურაციის ფაილის ჩაწერა შესაძლებელია. გირჩევთ წაშალოთ ფაილის (defines.php) ჩაწერის უფლებები.',
+            'The Monstra core directory (":path") and/or files underneath it has been found to be writable. We would advise you to remove all write permissions. <br>You can do this on unix systems with: <code>chmod -R a-w :path</code>' =>
+            'Monstra-ს დირექტორია (":path") ჩასაწერად ხელმისაწვდომია. გირჩევთ წაშალოთ ჩაწერის უფლებები დირექტორიაზე (":path") . <br>  ამის გაკეთება UNIX სისტემებზე ასე შეგიძლიათ: <code>chmod -R a-w :path</code>',
+            'The Monstra .htaccess file has been found to be writable. We would advise you to remove all write permissions. <br>You can do this on unix systems with: <code>chmod a-w :path</code>' =>
+            'მთავარი .htaccess ფაილი ჩასაწერად ხელმისაწვდომია. გირჩევთ წაშალოთ ჩაწერის უფლებები მთავარ .htaccess ფაილზე. <br> ამის გაკეთება UNIX სისტემებზე ასე შეგიძლიათ: <code>chmod -R a-w :path</code>',
+            'The Monstra index.php file has been found to be writable. We would advise you to remove all write permissions. <br>You can do this on unix systems with: <code>chmod a-w :path</code>' =>
+            'მთავარი index.php ფაილი ჩასაწერად ხელმისაწვდომია. გირჩევთ წაშალოთ ჩაწერის უფლებები მთავარ index.php ფაილზე. <br> ამის გაკეთება UNIX სისტემებზე ასე შეგიძლიათ: <code>chmod -R a-w :path</code>',
+            'Due to the type and amount of information an error might give intruders when Monstra::$environment = Monstra::DEVELOPMENT, we strongly advise setting Monstra::PRODUCTION in production systems.' =>
+            'სისტემა მუშაობს Monstra::DEVELOPMENT რეჟიმზა, გირჩევთ Monstra::PRODUCTION რეჟიმის დაყენებას.',
+            'Monstra version' => 'Monstra-ს ვერსია',
+            'Directory Permissions' => 'დაშვების უფლებები',
+            'Directory' => 'კატალოგი',
+            'Writable' => 'ჩასაწერად ხელმისაწვდომია',
+            'Unwritable' => 'ჩასაწერად არ არის ხელმისაწვდომია',
+            'Status' => 'სტატუსი',
+            'PHP Built On' => 'PHP აგებულია',
+            'Web Server' => 'Web სერვერი',
+            'WebServer to PHP Interface' => 'Web სერვერი PHP ინტერფეისისთვის',
+        )
+    );