+++ plugins/box/snippets/languages/ka-ge.lang.php
@@ -1,0 +1,30 @@
+<?php
+
+    return array(
+        'snippets' => array(
+            'Snippets' => 'Snippet-ები',
+            'Snippets manager' => 'Snippet-ების მენეჯერი',
+            'Actions' => 'მოქმედება',
+            'Delete' => 'წაშლა',
+            'Edit' => 'რედაქტირება',
+            'New Snippet' => 'ახალი Snippet-ი',
+            'Create New Snippet' => 'ახალი Snippet-ის შექმნა',
+            'Name' => 'სახელწოდება',
+            'Edit Snippet' => 'Snippet-ის რედაქტირება',
+            'Save' => 'შენახვა',
+            'Save and Exit' => 'შენახვა და შემდეგ გამოსვლა',
+            'This snippet already exists' => 'ასეთი Snippet-ი უკვე არსებობს',
+            'This snippet does not exist' => 'ასეთი Snippet-ი არ არსებობს',
+            'Delete snippet: :block' => 'Snippet-ის წაშლა: :snippet',
+            'Snippet content' => 'Snippet-ის შედგენილობა',
+            'Snippet <i>:name</i> deleted' => 'Snippet-ი <i>:name</i> წაშლილია',
+            'Your changes to the snippet <i>:name</i> have been saved.' => 'თქვენი ცვლილებები Snippet-ის მიმართ <i>:name</i> შენახულია.',
+            'Delete snippet: :snippet' => 'Snippet-ის წაშლა: :snippet',
+            'Required field' => 'აუცილებელი ველი',
+            'View Embed Code' => 'ჩასაწერი კოდი',
+            'Embed Code' => 'ჩასაწერი კოდი',
+            'Shortcode' => 'Shortcode-ი',
+            'PHP Code' => 'PHP კოდი',
+            'Cancel' => 'გაუქმება',
+        )
+    );