+++ plugins/markdown/markdown.plugin.php
@@ -12,17 +12,18 @@
  */
 
 // Register plugin
-Plugin::register( __FILE__,                    
+Plugin::register(__FILE__,
                 __('Markdown'),
-                __('Markdown markup language plugin for Monstra'),  
+                __('Markdown markup language plugin for Monstra'),
                 '1.0.0',
-                'Awilum',    
+                'Awilum',
                 'http://monstra.org/');
 
 // Uncomment code below to use Markdown on Site Content
 Filter::add('content', 'markdown', 1);
 
 use \Michelf\MarkdownExtra;
+
 include PLUGINS . '/markdown/php-markdown/Michelf/Markdown.php';
 include PLUGINS . '/markdown/php-markdown/Michelf/MarkdownExtra.php';