+++ plugins/sandbox/languages/ka-ge.lang.php
@@ -1,0 +1,10 @@
+<?php
+
+    return array(
+        'sandbox' => array(
+            'Sandbox' => 'Sandbox',
+            'Sandbox plugin for Monstra' => 'Sandbox-ის პლაგინი Monstra-თვის',
+            'Sandbox template' => 'Sandbox-ის ნიმუში',
+            'Save' => 'შენახვა',
+        )
+    );